seperated Magic numbers and added isAbsolute
