+++ bin/magic_numbers.dart
@@ -1,0 +1,43 @@
+import 'package:mime/mime.dart';
+
+MimeTypeResolver mimeTypeResolver() {
+  var resolver = MimeTypeResolver();
+
+  final magicNumber1 = [
+    0x00,
+    0x00,
+    0x00,
+    0x18,
+    0x66,
+    0x74,
+    0x79,
+    0x70,
+    0x4D,
+    0x34,
+    0x41
+  ];
+  final magicNumber2 = [
+    0x00,
+    0x00,
+    0x00,
+    0x14,
+    0x66,
+    0x74,
+    0x79,
+    0x70,
+    0x4D,
+    0x34,
+    0x41
+  ];
+
+  resolver.addMagicNumber(magicNumber1, 'audio/mp4');
+  resolver.addMagicNumber(magicNumber2, 'audio/mp4');
+  resolver.addMagicNumber(
+      [0x00, 0x00, 0x00, 0x18, 0x66, 0x74, 0x79, 0x70, 0x4D, 0x34, 0x41],
+      'audio/mp4');
+  resolver
+      .addMagicNumber([0x66, 0x74, 0x79, 0x70, 0x4D, 0x34, 0x41], 'audio/mp4');
+  resolver.addMagicNumber(
+      [0x66, 0x74, 0x79, 0x70, 0x69, 0x73, 0x6F, 0x6D], 'audio/mp4');
+  return resolver;
+}
