+++ bin/mime_sorter.dart
@@ -5,10 +5,11 @@ import 'dart:convert';
 import 'dart:io';
 import 'package:dcli/dcli.dart';
 import 'package:args/args.dart';
-import 'package:mime/mime.dart';
+import './magic_numbers.dart';
 import 'package:path/path.dart' as p;
 
 const String version = '0.0.1';
+var resolver = mimeTypeResolver();
 
 ArgParser buildParser() {
   return ArgParser()
@@ -41,45 +42,6 @@ void printUsage(ArgParser argParser) {
 }
 
 void main(List<String> arguments) {
-  var resolver = MimeTypeResolver();
-
-  final magicNumber1 = [
-    0x00,
-    0x00,
-    0x00,
-    0x18,
-    0x66,
-    0x74,
-    0x79,
-    0x70,
-    0x4D,
-    0x34,
-    0x41
-  ];
-  final magicNumber2 = [
-    0x00,
-    0x00,
-    0x00,
-    0x14,
-    0x66,
-    0x74,
-    0x79,
-    0x70,
-    0x4D,
-    0x34,
-    0x41
-  ];
-
-  resolver.addMagicNumber(magicNumber1, 'audio/mp4');
-  resolver.addMagicNumber(magicNumber2, 'audio/mp4');
-  resolver.addMagicNumber(
-      [0x00, 0x00, 0x00, 0x18, 0x66, 0x74, 0x79, 0x70, 0x4D, 0x34, 0x41],
-      'audio/mp4');
-  resolver
-      .addMagicNumber([0x66, 0x74, 0x79, 0x70, 0x4D, 0x34, 0x41], 'audio/mp4');
-  resolver.addMagicNumber(
-      [0x66, 0x74, 0x79, 0x70, 0x69, 0x73, 0x6F, 0x6D], 'audio/mp4');
-
   final ArgParser argParser = buildParser();
   try {
     final ArgResults results = argParser.parse(arguments);
@@ -99,6 +61,9 @@ void main(List<String> arguments) {
     }
     var targetDirectory =
         results.option('source_Directory') ?? Directory.current.path;
+    targetDirectory = p.isAbsolute(targetDirectory)
+        ? targetDirectory
+        : p.absolute(targetDirectory);
     mimeTyper(targetDirectory);
 
     // Act on the arguments provided.
@@ -115,10 +80,10 @@ void main(List<String> arguments) {
 }
 
 void mimeTyper(String targetDirectory) async {
-  checkPath(targetDirectory);
+  targetDirectory = checkPath(targetDirectory);
   print("We proceeded");
-
-  var files = Directory(p.normalize(targetDirectory))
+  print(targetDirectory);
+  /*  var files = Directory(p.normalize(targetDirectory))
       .list(recursive: true, followLinks: true);
 
   await for (var fileToSort in files) {
@@ -126,16 +91,16 @@ void mimeTyper(String targetDirectory) async {
     fileToSort is Directory
         ? print("Skipping $fileToSort because it is a Directory")
         : moveFiles(fileToSort, targetDirectory);
-  }
+  } */
 }
 
-checkPath(String targetDirectory) {
-  targetDirectory = p.isAbsolute(targetDirectory)
-      ? targetDirectory
-      : p.absolute(targetDirectory);
+String checkPath(String targetDirectory) {
+  targetDirectory = p.absolute(targetDirectory);
+  targetDirectory = p.canonicalize(targetDirectory);
   var proceed =
       confirm("Did you want to use $targetDirectory", defaultValue: false);
   proceed == true ? "" : exit(0);
+  return targetDirectory;
 }
 
 moveFile(FileSystemEntity targetFile, String targetDirectory) async {
@@ -197,7 +162,7 @@ Future<String> findMIMEType(FileSystemEntity targetFile) async {
 
 logFiles(FileSystemEntity targetFile) async {
   var filePath = targetFile.path;
-  var dirName = lookupMimeType(filePath,
+  var dirName = resolver.lookup(filePath,
           headerBytes: await File(filePath).openRead(0, 25).first) ??
       'null';
   dirName.contains('audio') ? print(' : $dirName') : null;
@@ -206,7 +171,7 @@ logFiles(FileSystemEntity targetFile) async {
 
 moveFiles(FileSystemEntity targetFile, String targetDirectory) async {
   var filePath = targetFile.path;
-  var dirName = lookupMimeType(filePath,
+  var dirName = resolver.lookup(filePath,
           headerBytes: await File(filePath).openRead(0, 25).first) ??
       'null';
   var sortedDir = dirName != "null" ? dirName.split("/").first : 'Unknown';
